Work around gcc 4.8.0 issue on ppc64
Change iteration to use size_t instead of int.  The code should be
equivalent in all reasonable cases, but the original way this was
coded was causing a test failure with gcc 4.8.0 on ppc64.  See
https://bugzilla.redhat.com/show_bug.cgi?id=915321 for additional
information.

+++ qpdf/qpdf.cc
@@ -1617,7 +1617,7 @@ int main(int argc, char* argv[])
                 // This prevents those objects from being preserved by
                 // being referred to from other places, such as the
                 // outlines dictionary.
-                for (int pageno = 0; pageno < (int)orig_pages.size(); ++pageno)
+                for (size_t pageno = 0; pageno < orig_pages.size(); ++pageno)
                 {
                     if (selected_from_orig.count(pageno) == 0)
                     {
